Zhuoran Wu: change styles of file structures
