+++ input/clean/earning_info_CLEANED.csv
