+++ input/raw/earning_info.csv
