--- src/README.txt
@@ -1,41 +1,0 @@
-Python files:
-
-1) gdp_info_CLEANING.py: cleans the gdp_info.csv and outputs the cleaned data 
-			in gdp_info_CLEANED.csv.
-
-2) crimeRatesCLEANING.py: cleans the crime_rates.csv and outputs the cleaned data
-			  in crime_rates_CLEANED.csv.
-
-3) createZillowOutput.py: merges Zillow_House_Update_40000.csv, Zillow_House_Info_40000.csv,
-			  and Zillow_House_Detail_40000.csv into one csv: Zillow_House_Output.csv.
-
-4) cleanZillowOutput.py: cleans the Zillow_House_Output.csv and outputs the cleaned data
-			 in Zillow_Cleaned.csv.
-
-5) cleanGraduationRates.py: cleans the graduation_rates.csv and outputs the cleaned data
-			    in graduation_rates_CLEANED.csv.
-
-6) cleanEarningInfo.py: cleans the earning_info.csv and outputs the cleaned data
-			in earning_info_CLEANED.csv.
-
-7) cleanCrimeCounts.py: cleans the crime_counts.csv and outputs the cleaned data
-			in crime_counts_CLEANED.csv.
-
-
-CSV files:
-
-1) Zillow_House_Update_40000: Contains information about the house amenities.
-2) Zillow_House_Info_40000: Contains information about the house's location and price.
-3) Zillow_House_Detail_40000: Contains information about the house's rooms, year built, etc.
-4) Zillow_House_Output.csv: The output csv file after merging files (1) - (3).
-5) Zillow_Cleaned.csv: The cleaned csv file for Zillow_House_Output.csv.
-6) graduation_rates.csv: Contains information about graduation rates for a given area.
-7) graduation_rates_CLEANED.csv: The cleaned csv file for graduation_rates.csv.
-8) gdp_info.csv: Contains gdp info for a given area.
-9) gdp_info_CLEANED.csv: The cleaned csv file for gdp_info.csv.
-10) earning_info.csv: Contains info about earnings for a given area.
-11) earning_info_CLEANED.csv: The cleaned csv file for earning_info.csv.
-12) crime_rates.csv: Contains info about crime rates for a given area.
-13) crime_rates_CLEANED.csv: The cleaned csv file for crime_rates.csv.
-14) crime_counts.csv: Contains info about the crime count for a given area.
-15) crime_counts_CLEANED.csv: The cleaned csv file for crime_counts.csv.
