+++ src/data_cleaning/cleanEarningInfo.py
@@ -1,4 +1,4 @@
-# -*- coding: utf-8 -*-
+#  -*- coding: utf-8 -*-
 """
 Created on Thu Oct  5 14:52:34 2017
 
@@ -12,28 +12,29 @@ the row.
 import pandas as pd
 import sys
 
+
 def main():
-    df = pd.read_csv("earning_info.csv") #read in the input file into the dataframe.
-    df = df.dropna() #dropping the rows with missing observations.
+    df = pd.read_csv("../../input/raw/earning_info.csv") # read in the input file into the dataframe.
+    df = df.dropna() # dropping the rows with missing observations.
     
-    #Dropping all rows with any negative values.
+    # Dropping all rows with any negative values.
     df = df[df>0].dropna() 
     
-    #Dropping rows where the year is less than 2000.
+    # Dropping rows where the year is less than 2000.
     df = df.drop(df[(df["year"]) < 2000].index)
     
-    #Dropping rows where the year greater than 2017, the current year.
+    # Dropping rows where the year greater than 2017, the current year.
     df = df.drop(df[(df["year"]) > 2017].index)
     
-    #Dropping all rows where the percentages of categories of earnings don't add up to 100.
+    # Dropping all rows where the percentages of categories of earnings don't add up to 100.
     df = df.drop(df[100 != (df["percent_with_earnings_10000_to_14999"] + df["percent_with_earnings_15000_to_24999"] + df["percent_with_earnings_1_to_9999"] + df["percent_with_earnings_25000_to_34999"] + df["percent_with_earnings_35000_to_49999"] + df["percent_with_earnings_50000_to_64999"] + df["percent_with_earnings_65000_to_74999"] + df["percent_with_earnings_75000_to_99999"] + df["percent_with_earnings_over_100000"]  )].index)
     
-    #Splitting the area_name column into City and State.
-    df["City"], df["State"] = zip(*df["area_name"].str.split("- ").tolist())     #Splitting the area_name column into City and State.
-    del df["area_name"] #now city and state are separated so we get rid of "City, State"
+    # Splitting the area_name column into City and State.
+    df["City"], df["State"] = zip(*df["area_name"].str.split("- ").tolist())     # Splitting the area_name column into City and State.
+    del df["area_name"] # now city and state are separated so we get rid of "City, State"
     
-    #Finally outputting the dataframe to a csv file.
-    df.to_csv("earning_info_CLEANED.csv")
+    # Finally outputting the dataframe to a csv file.
+    df.to_csv("../../input/earning_info_CLEANED.csv")
 
 if __name__ == "__main__":
     main()
